Changed the project's directory structure so its easier to understand the different components
