+++ src/simulator/player-actions/PlayerActions.ts
@@ -1,0 +1,21 @@
+import { Indicators, WorldState } from '../SimulatorState';
+
+/**
+ * Containment policies are specific actions that can be taken in conjunction, in the context of a containment doctrine,
+ * assuming the required capability improvement requirements are met.
+ */
+export interface ContainmentPolicy {
+    requirements: CapabilityImprovements[];
+    name: string;
+    immediateEffect: (context: WorldState) => Indicators;
+    recurringEffect: (context: WorldState) => Indicators;
+}
+
+/**
+ * Represents specific improvements to the healthcare sector (e.g. contact tracing infrastructure)
+ */
+export interface CapabilityImprovements {
+    name: string;
+    immediateEffect: (context: WorldState) => Indicators;
+    recurringEffect: (context: WorldState) => Indicators;
+}
