+++ src/simulator/victory-conditions/VictoryConditon.ts
@@ -1,0 +1,7 @@
+import { SimulatorState } from '../SimulatorState';
+
+export interface VictoryCondition {
+    name: string;
+    description: string;
+    isMet(simulatorState: SimulatorState): boolean;
+}
